Refactor workflow and session handling; introduce batch action orchestration
- Replaced direct JSON writing with atomic write operations in `workflow.py` and `session.py` for improved safety.
- Extracted batch action request orchestration into a new module `agent_batch.py`, isolating the batch workflow logic.
- Added telemetry content-safety helpers in `telemetry_safety.py` for redaction and validation of sensitive data.
- Implemented unit tests for command templates in `test_command_templates.py` to ensure proper quoting and command generation.
