+++ .github/workflows/ci.yml
@@ -27,13 +27,17 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Install lint dependencies
-        run: python -m pip install --upgrade pip ruff build
+        run: python -m pip install --upgrade pip ruff build coverage
 
       - name: Ruff
         run: ruff check src tests
 
-      - name: Unit tests
-        run: python -m unittest discover -s tests
+      - name: Unit tests (with coverage)
+        env:
+          PYTHONPATH: src
+        run: |
+          python -m coverage run -m unittest discover -s tests
+          python -m coverage report
 
       - name: Generate plugin payload
         run: python scripts/build_plugin_payload.py --output dist/plugin/gh-address-cr
