+++ .gitignore
@@ -8,6 +8,9 @@ dist/
 build/
 *.egg-info/
 .coverage
+.coverage.*
+htmlcov/
+coverage.xml
 coverage/
 *.tmp
 *.log
