+++ src/gh_address_cr/core/session.py
@@ -103,11 +103,9 @@ def load_session(repo: str, pr_number: str) -> dict[str, Any]:
     payload.setdefault("leases", {})
     payload.setdefault("ledger_path", str(default_ledger_path(repo, pr_number)))
     _coerce_lease_datetimes(payload)
-    try:
-        from gh_address_cr.core.telemetry import SessionTelemetry
-        SessionTelemetry.get_instance().configure_context(repo, str(pr_number))
-    except Exception:
-        pass
+    from gh_address_cr.core.telemetry import configure_context_safely
+
+    configure_context_safely(repo, pr_number)
     return payload
